6.2 completed register screen
